+++ Modules/Microsoft365DSC/Modules/M365DSCUtil.psm1
@@ -1487,7 +1487,7 @@ function Export-M365DSCConfiguration
     Add-M365DSCTelemetryEvent -Type 'ExportInitiated' -Data $data
     if ($null -ne $Workloads)
     {
-        Write-Output -InputObject "Exporting Microsoft 365 configuration for Workloads: $($Workloads -join ', ')"
+        Write-M365DSCHost -Message "Exporting Microsoft 365 configuration for Workloads: $($Workloads -join ', ')"
         Start-M365DSCConfigurationExtract -Credential $Credential `
             -Workloads $Workloads `
             -Mode $Mode `
@@ -1508,7 +1508,7 @@ function Export-M365DSCConfiguration
     }
     elseif ($null -ne $Components)
     {
-        Write-Output -InputObject "Exporting Microsoft 365 configuration for Components: $($Components -join ', ')"
+        Write-M365DSCHost -Message "Exporting Microsoft 365 configuration for Components: $($Components -join ', ')"
         Start-M365DSCConfigurationExtract -Credential $Credential `
             -Components $Components `
             -Path $Path -FileName $FileName `
@@ -1528,7 +1528,7 @@ function Export-M365DSCConfiguration
     }
     elseif ($null -ne $Mode)
     {
-        Write-Output -InputObject "Exporting Microsoft 365 configuration for Mode: $Mode"
+        Write-M365DSCHost -Message "Exporting Microsoft 365 configuration for Mode: $Mode"
         Start-M365DSCConfigurationExtract -Credential $Credential `
             -Mode $Mode `
             -Path $Path -FileName $FileName `
@@ -5136,18 +5136,18 @@ function Get-M365DSCConfigurationConflict
 }
 
 <#
-        .Description
-        This function returns a hashtable with aligned to the parameter pattern of the given cmdlet.
+.Description
+This function returns a hashtable with aligned to the parameter pattern of the given cmdlet.
 
-        .Example
-        $param = @{
-            Path = 'C:\Test'
-            DoesNotExist = '123'
-        }
-        Sync-M365DSCParameter -Command (Get-Command -Name Get-ChildItem) -Parameters $param
+.Example
+$param = @{
+    Path = 'C:\Test'
+    DoesNotExist = '123'
+}
+Sync-M365DSCParameter -Command (Get-Command -Name Get-ChildItem) -Parameters $param
 
-        .Functionality
-        Private
+.Functionality
+Private
 #>
 function Sync-M365DSCParameter
 {
@@ -5264,17 +5264,22 @@ function Sync-M365DSCParameter
 <#
 .SYNOPSIS
     Joins two or more M365DSC configurations into a single configuration.
+
 .DESCRIPTION
     This function is used to join two or more M365DSC configurations into a single configuration.
     The function reads the configuration from the specified paths and combines them into a single configuration.
     Please note that the function won't be updating the authentication parameters if they differ between the configurations. Make sure that the authentication parameters are the same over all configurations.
+
 .PARAMETER ConfigurationFile
     The name of the first configuration file to use as the base configuration.
+
 .PARAMETER ConfigurationPath
     The directory path to the configuration files to join to the base configuration.
+
 .EXAMPLE
     Join-M365DSCConfiguration -ConfigurationFile 'M365TenantConfig.ps1' -ConfigurationPath 'D:\testbed'
     This example joins the 'M365TenantConfig.ps1' file with all the configuration files in the 'D:\testbed' directory.
+
 .FUNCTIONALITY
     Public
 #>
@@ -5350,7 +5355,8 @@ function Join-M365DSCConfiguration
 .OUTPUTS
     Result of the invoked function.
 #>
-function Invoke-PowerShellCoreResource {
+function Invoke-PowerShellCoreResource
+{
     [CmdletBinding()]
     param (
         [Parameter(Mandatory = $true)]
@@ -5390,7 +5396,8 @@ function Invoke-PowerShellCoreResource {
 .EXAMPLE
     Initialize-PowerShellCoreSession
 #>
-function Initialize-PowerShellCoreSession {
+function Initialize-PowerShellCoreSession
+{
     $script:PSCoreSession = New-PSSession -ComputerName localhost -ConfigurationName PowerShell.7 -EnableNetworkAccess
     $lcmConfig = Get-DscLocalConfigurationManager
     Invoke-Command -Session $script:PSCoreSession -ScriptBlock {
@@ -5399,6 +5406,88 @@ function Initialize-PowerShellCoreSession {
         Set-M365DSCLCMConfiguration -LCMConfig $using:lcmConfig
     }
     $script:PSCoreSessionInitialized = $true
+}
+
+<#
+.Description
+This function writes messages to the console or verbose output.
+
+.PARAMETER Message
+Specifies the message to write.
+
+.PARAMETER DeferWrite
+Specifies if writing the message should be deferred. Adheres to -NoNewLine behavior of Write-Host.
+
+.PARAMETER CommitWrite
+Specifies if cached messages of -DeferWrite should be combined and written.
+Combining of the messages is done by joining them without any characters between.
+
+.EXAMPLE
+Write-M365DSCHost -Message "This is a message."
+
+.Functionality
+Internal
+#>
+function Write-M365DSCHost
+{
+    [CmdletBinding(DefaultParameterSetName = 'Default')]
+    param
+    (
+        [Parameter(Mandatory = $true, Position = 0)]
+        [System.String]
+        $Message,
+
+        [Parameter()]
+        [ConsoleColor]
+        $ForegroundColor = [System.Console]::ForegroundColor,
+
+        [Parameter(ParameterSetName = 'DeferWrite')]
+        [switch]
+        $DeferWrite,
+
+        [Parameter(ParameterSetName = 'CommitWrite')]
+        [switch]
+        $CommitWrite
+    )
+
+    if ($null -eq $Script:M365DSCHostMessages)
+    {
+        $Script:M365DSCHostMessages = @()
+    }
+
+    if ($DeferWrite)
+    {
+        $Script:M365DSCHostMessages += @{
+            Message = $Message
+            ForegroundColor = $ForegroundColor
+        }
+        return
+    }
+
+    if ([Environment]::UserInteractive)
+    {
+        if ($CommitWrite -and $Script:M365DSCHostMessages.Count -gt 0)
+        {
+            for ($i = 0; $i -lt $Script:M365DSCHostMessages.Count - 1; $i++)
+            {
+                Write-Host -Object $Script:M365DSCHostMessages[$i].Message -ForegroundColor $Script:M365DSCHostMessages[$i].ForegroundColor -NoNewline
+            }
+            Write-Host -Object $Script:M365DSCHostMessages[-1].Message -ForegroundColor $Script:M365DSCHostMessages[-1].ForegroundColor -NoNewline
+            $Script:M365DSCHostMessages = @()
+        }
+        Write-Host -Object $Message -ForegroundColor $ForegroundColor
+    }
+    else
+    {
+        $outputMessage = ''
+        if ($CommitWrite)
+        {
+            $outputMessage += $Script:M365DSCHostMessages.Message -join ''
+            $Script:M365DSCHostMessages = @()
+        }
+        $finalMessage = $outputMessage + $Message
+        Write-Verbose -Message $finalMessage -Verbose
+    }
 }
 
 Export-ModuleMember -Function @(
@@ -5449,5 +5538,6 @@ Export-ModuleMember -Function @(
     'Update-M365DSCModule',
     'Write-M365DSCLogEvent',
     'Sync-M365DSCParameter',
-    'Invoke-PowerShellCoreResource'
+    'Invoke-PowerShellCoreResource',
+    'Write-M365DSCHost'
 )